requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-sse-client==0.2.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.1
attrs==24.2.0
bitarray==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
python-dotenv==1.0.1
pytonapi==0.4.1
pytonconnect==0.3.1
redis==5.1.1
requests==2.32.3
sniffio==1.3.1
starlette==0.39.2
tonapi==0.1.0
tonsdk==1.0.15
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
websockets==13.1
yarl==1.15.5
